rest_api/flask_server/requirements_pinned_versions.txt
Flask==2.0.1
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.1
pybullet==3.1.7
pyrealsense2==2.49.0.3474
roslibpy==1.2.1
toml==0.10.2
